[version : v0.0.1] [srv-crc16 : ACEC] 分离sdk中间件对象，管理，中间件处理逻辑
