+++ pack/release_version.json
@@ -5,7 +5,7 @@
             "0.0.0"
         ],
         [
-            "Sun Feb 16 16:28:54 2020",
+            "Sun Feb 16 17:20:46 2020",
             "0.0.1"
         ]
     ]
