+++ src/base_fun/sdk_net/sdk_network/Makefile
@@ -1,4 +1,7 @@
-OBJS := sdk_network.o sdk_protocol_do.o sdk_midware.o sdk_interface.o
+OBJS := sdk_network.o \
+		sdk_protocol_do.o \
+		sdk_midware.o sdk_midware_manage.o sdk_midware_do.o \
+		sdk_interface.o
 DIRS := 
 
 include $(SUB_MAKE_INCLUDE)
